# music/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music API",
        "description": "Documentation de l'API Musicbox",
        "contact": {
            "name": "Ayaah Mhaouar",
            "email": "ayaah@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/music": {
            "get": {
                "tags": [
                    "Music"
                ],
                "summary": "Liste de toutes les musiques",
                "operationId": "6aa33eaffdc3057bfa830b63dc95f3e8",
                "responses": {
                    "200": {
                        "description": "Liste de musiques",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Song 1"
                                            },
                                            "artist": {
                                                "type": "string",
                                                "example": "Artist 1"
                                            },
                                            "album": {
                                                "type": "string",
                                                "example": "Album 1"
                                            },
                                            "duration": {
                                                "type": "integer",
                                                "example": 180
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Music"
                ],
                "summary": "Créer une nouvelle musique",
                "operationId": "15ee08e9519c2d8e749ea57ec8dfff14",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "artist"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Song 1"
                                    },
                                    "artist": {
                                        "type": "string",
                                        "example": "Artist 1"
                                    },
                                    "album": {
                                        "type": "string",
                                        "example": "Album 1"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 180
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Musique créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Song 1"
                                        },
                                        "artist": {
                                            "type": "string",
                                            "example": "Artist 1"
                                        },
                                        "album": {
                                            "type": "string",
                                            "example": "Album 1"
                                        },
                                        "duration": {
                                            "type": "integer",
                                            "example": 180
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/music/{id}": {
            "get": {
                "tags": [
                    "Music"
                ],
                "summary": "Voir une musique avec sa playlist",
                "operationId": "c62b53bb526e5c9c2a9993f8b5165ac4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Musique trouvée avec sa playlist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Lo-Fi Beat"
                                        },
                                        "artist": {
                                            "type": "string",
                                            "example": "DJ Relax"
                                        },
                                        "album": {
                                            "type": "string",
                                            "example": "Lo-Fi Collection"
                                        },
                                        "duration": {
                                            "type": "integer",
                                            "example": 180
                                        },
                                        "playlist": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Chill Vibes"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Music"
                ],
                "summary": "Mettre à jour une musique",
                "operationId": "8b06251ff3375ae8243be8b9200a3a36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Song"
                                    },
                                    "artist": {
                                        "type": "string",
                                        "example": "Updated Artist"
                                    },
                                    "album": {
                                        "type": "string",
                                        "example": "Updated Album"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 200
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Musique mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Updated Song"
                                        },
                                        "artist": {
                                            "type": "string",
                                            "example": "Updated Artist"
                                        },
                                        "album": {
                                            "type": "string",
                                            "example": "Updated Album"
                                        },
                                        "duration": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Music"
                ],
                "summary": "Supprimer une musique",
                "operationId": "ed7584c7dc0e7de0b6bea1e26433c10d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Musique supprimée"
                    }
                }
            }
        },
...
    },
...
}